fix: Move FolderPicker
